⚡ separated labels variables and relationships in constants
