+++ libs/api/database/neo4j/src/lib/constants/repository/repository.labels.ts
@@ -1,0 +1,9 @@
+enum LABEL {
+  Address = 'ADDRESS',
+  AddressOf = 'ADDRESS_OF',
+  Person = 'PERSON',
+  Trip = 'TRIP',
+  AccompliceOf = 'ACCOMPLICE_OF'
+}
+
+export { LABEL };
