+++ libs/api/database/neo4j/src/lib/constants/repository/repository.variables.ts
@@ -1,0 +1,8 @@
+enum VARIABLE {
+  Address = 'address',
+  Person = 'person',
+  Trip = 'trip',
+  Output = 'output'
+}
+
+export { VARIABLE };
